requirements.txt
asgiref==3.5.2
boto3==1.24.10
botocore==1.27.10
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.51.0
django-appconf==1.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-imagekit==4.1.0
django-rest-auth==0.9.5
django-sslserver==0.22
django-storages==1.12.3
djangorestframework==3.13.1
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
jmespath==1.0.0
msgpack==1.0.4
oauthlib==3.2.0
pilkit==2.0
Pillow==9.1.1
proto-plus==1.20.6
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyheif==0.7.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whatimage==0.0.3
